feature_eng/pyproject.toml
[tool.poetry]
name = "feature-eng"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
sentence-transformers = "^3.3.1"
openai = "^1.56.1"
seaborn = "^0.13.2"
pubchempy = "^1.0.4"
bs4 = "^0.0.2"
biopython = "^1.84"
rdkit = "^2024.3.6"
tiktoken = "^0.8.0"
anthropic = "^0.42.0"
google-genai = "^0.3.0"
nltk = "^3.9.1"
google-generativeai = "^0.8.4"
adjusttext = "^1.3.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
